requirements.txt
# Core dependencies
numpy>=2.2.0,<2.3.0  # Upper bound for numba compatibility
scipy>=1.14.0
pandas>=2.2.0
scikit-learn>=1.5.0
torch>=2.8.0
torchvision>=0.21.0

# Protein language models
fair-esm==2.0.0
transformers>=4.57.0
biopython>=1.84
huggingface-hub>=0.28.0
sentencepiece>=0.2.1

# Data processing
h5py>=3.12.0
einops>=0.8.0
pyyaml>=6.0
tqdm>=4.67.0
datasets>=3.2.0
multiprocess>=0.70.0

# Analysis and training
nnsight>=0.5.7
typed-argument-parser>=1.10.0
umap-learn>=0.5.8

# Visualization and dashboard
streamlit>=1.41.0
plotly>=5.24.0
matplotlib>=3.10.0
seaborn>=0.13.0
py3dmol>=2.4.0

# Experiment tracking
wandb>=0.19.0